refactor: preparation for adding other frameworks
