+++ src/frameworks/mod.rs
@@ -1,0 +1,5 @@
+/// Defines the `Framework` trait.
+pub mod framework;
+
+/// Defines the `Foundry` struct which implements the `Framework` trait.
+pub mod foundry;
